refactor exception, look at security, change db schema for classes
