+++ src/main/resources/db/migration/V4__rectify_class_course_nullable.sql
@@ -1,0 +1,4 @@
+SET search_path TO lms;
+
+ALTER TABLE classes
+ALTER COLUMN course_id DROP NOT NULL;
